+++ src/comm/ingest.rs
@@ -58,7 +58,7 @@ const CHANNEL_CLOSE_MESSAGE: &[u8; 12] = b"channel done";
 const CHANNEL_CLOSE_TIMESTAMP: i64 = -1;
 const NO_TIMESTAMP: i64 = 0;
 const SENSOR_INTERVAL: u64 = 60 * 60 * 24;
-const INGEST_VERSION_REQ: &str = ">=0.26.0-alpha.6,<0.26.0-alpha.7";
+const INGEST_VERSION_REQ: &str = ">=0.26.0-alpha.6,<0.26.0-alpha.8";
 
 type SensorInfo = (String, DateTime<Utc>, ConnState, bool);
 
@@ -266,6 +266,20 @@ async fn handle_request(
                 Some(NetworkKey::new(&sensor, "dns")),
                 sensor,
                 db.dns_store()?,
+                stream_direct_channels,
+                shutdown_signal,
+                ack_trans_cnt,
+            )
+            .await?;
+        }
+        RawEventKind::MalformedDns => {
+            handle_data(
+                send,
+                recv,
+                RawEventKind::MalformedDns,
+                Some(NetworkKey::new(&sensor, "malformed_dns")),
+                sensor,
+                db.malformed_dns_store()?,
                 stream_direct_channels,
                 shutdown_signal,
                 ack_trans_cnt,